fix: update tests and extract internal arrow function components
- Expand internalArrowFunctionDeclarations in parseReactComponents.js to include components with the ArrowFunctionExpression init_type
- Remove obsolete components.test.js file
- Move the component nested in factory test from dataExtraction to edgeCases test suite
- Update edgeCases tests to use new parsing modules instead of parseCode
